pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-search-hub"
version = "0.1.0"
description = "FastMCP 2.0 server integrating multiple search providers"
authors = [{ name = "Bjorn Melin" }]
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.3.4",
    "pydantic>=2.11.0",
    "httpx>=0.28.1",
    "uvicorn>=0.34.2",
    "python-dotenv>=1.1.0",
]

[project.scripts]
mcp-search-hub = "mcp_search_hub.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
]

[tool.setuptools]
packages = ["mcp_search_hub"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "C",    # flake8-comprehensions
    "B",    # flake8-bugbear
    "A",    # flake8-builtins
    "SIM",  # flake8-simplify
    "PIE",  # flake8-pie
    "FLY",  # flynt
    "PL",   # pylint
    "RET",  # flake8-return
    "ANN",  # flake8-annotations
    "ASYNC", # flake8-async
    "LOG",  # flake8-logging
]
ignore = [
    "E501",  # line too long (handled by black)
    "ANN101", # self annotations
    "ANN102", # cls annotations
    "PLR0913", # Too many arguments
]
